:hammer: Move the PEP class under app data rather than API
With an eye on being able to handle location annotations and checking if a
PEP has been changed or is new, etc, here I'm moving the PEP class out of
the API code and under the app code.

From now on the API code simply returns back the raw data, it's then up to
the application to turn that into something more organised. This means I can
safely have extra properties that come from storage.
