+++ src/peplum/app/data/__init__.py
@@ -8,6 +8,7 @@ from .config import (
     save_configuration,
     update_configuration,
 )
+from .pep import PEP, PEPStatus, PEPType, PostHistory
 from .peps import (
     AuthorCount,
     Containing,
@@ -31,8 +32,12 @@ __all__ = [
     "Configuration",
     "Containing",
     "load_configuration",
+    "PEP",
+    "PEPStatus",
+    "PEPType",
     "PEPCount",
     "PEPs",
+    "PostHistory",
     "PythonVersionCount",
     "save_configuration",
     "StatusCount",
